Split test file into two files
