+++ tests/graph/test_centrality.py
@@ -15,37 +15,6 @@ import unittest
 import retworkx
 
 
-class TestCentralityDiGraph(unittest.TestCase):
-    def setUp(self):
-        self.graph = retworkx.PyDiGraph()
-        self.a = self.graph.add_node("A")
-        self.b = self.graph.add_node("B")
-        self.c = self.graph.add_node("C")
-        self.d = self.graph.add_node("D")
-        edge_list = [
-            (self.a, self.b, 1),
-            (self.b, self.c, 1),
-            (self.c, self.d, 1),
-        ]
-        self.graph.add_edges_from(edge_list)
-
-    def test_betweenness_centrality(self):
-        betweenness = retworkx.digraph_betweenness_centrality(self.graph)
-        expected = {0: 0.0, 1: 0.3333333333333333, 2: 0.3333333333333333, 3: 0.0}
-        self.assertEqual(expected, betweenness)
-
-    def test_betweenness_centrality_endpoints(self):
-        betweenness = retworkx.digraph_betweenness_centrality(
-            self.graph, endpoints=True)
-        expected = {0: 0.25, 1: 0.41666666666666663, 2: 0.41666666666666663, 3: 0.25}
-        self.assertEqual(expected, betweenness)
-
-    def test_betweenness_centrality_unnormalized(self):
-        betweenness = retworkx.digraph_betweenness_centrality(
-            self.graph, endpoints=False, normalized=False)
-        expected = {0: 0.0, 1: 2.0, 2: 2.0, 3: 0.0}
-        self.assertEqual(expected, betweenness)
-
 
 class TestCentralityGraph(unittest.TestCase):
     def setUp(self):
@@ -79,9 +48,9 @@ class TestCentralityGraph(unittest.TestCase):
         self.assertEqual(expected, betweenness)
 
 
-class TestCentralityDiGraphDeletedNode(unittest.TestCase):
+class TestCentralityGraphDeletedNode(unittest.TestCase):
     def setUp(self):
-        self.graph = retworkx.PyDiGraph()
+        self.graph = retworkx.PyGraph()
         self.a = self.graph.add_node("A")
         self.b = self.graph.add_node("B")
         self.c = self.graph.add_node("C")
@@ -96,18 +65,18 @@ class TestCentralityDiGraphDeletedNode(unittest.TestCase):
         self.graph.remove_node(c0)
 
     def test_betweenness_centrality(self):
-        betweenness = retworkx.digraph_betweenness_centrality(self.graph)
-        expected = {0: 0.0, 1: 0.3333333333333333, 2: 0.3333333333333333, 4: 0.0}
+        betweenness = retworkx.graph_betweenness_centrality(self.graph)
+        expected = {0: 0.0, 1: 0.6666666666666666, 2: 0.6666666666666666, 4: 0.0}
         self.assertEqual(expected, betweenness)
 
     def test_betweenness_centrality_endpoints(self):
-        betweenness = retworkx.digraph_betweenness_centrality(
+        betweenness = retworkx.graph_betweenness_centrality(
             self.graph, endpoints=True)
-        expected = {0: 0.25, 1: 0.41666666666666663, 2: 0.41666666666666663, 4: 0.25}
+        expected = {0: 0.5, 1: 0.8333333333333333, 2: 0.8333333333333333, 4: 0.5}
         self.assertEqual(expected, betweenness)
 
     def test_betweenness_centrality_unnormalized(self):
-        betweenness = retworkx.digraph_betweenness_centrality(
+        betweenness = retworkx.graph_betweenness_centrality(
             self.graph, endpoints=False, normalized=False)
         expected = {0: 0.0, 1: 2.0, 2: 2.0, 4: 0.0}
         self.assertEqual(expected, betweenness)
